Centralized get_gre_date(..) => less code duplication, better test coverage
