+++ holidays/utils.py
@@ -1,6 +1,6 @@
 import inspect
-
 import holidays.countries as countries
+from datetime import date
 
 
 def list_supported_countries():
@@ -19,3 +19,37 @@ def CountryHoliday(country, years=[], prov=None, state=None, expand=True,
     except StopIteration:
         raise KeyError("Country %s not available" % country)
     return country_holiday
+
+
+def get_gre_date(year, Hmonth, Hday):
+    """
+    returns the gregian date of of a  of the given gregorian calendar
+    yyyy year with Hijari Month & Day
+    works *only* if hijri-converter library is installed, otherwise a warning
+    is raised that this holiday is missing. hijri-converter requires
+    Python >= 3.6
+    """
+    try:
+        from hijri_converter import convert
+    except ImportError:
+        import warnings
+
+        def warning_on_one_line(message, category, filename, lineno,
+                                file=None, line=None):
+            return filename + ': ' + str(message) + '\n'
+        warnings.formatwarning = warning_on_one_line
+        warnings.warn("Error estimating Islamic Holidays." +
+                      "To estimate, install hijri-converter library")
+        warnings.warn("pip install -U hijri-converter")
+        warnings.warn("(see https://hijri-converter.readthedocs.io/ )")
+        return []
+    Hyear = convert.Gregorian(year, 1, 1).to_hijri().datetuple()[0]
+    gres = []
+    gres.append(convert.Hijri(Hyear - 1, Hmonth, Hday).to_gregorian())
+    gres.append(convert.Hijri(Hyear, Hmonth, Hday).to_gregorian())
+    gres.append(convert.Hijri(Hyear + 1, Hmonth, Hday).to_gregorian())
+    gre_dates = []
+    for gre in gres:
+        if gre.year == year:
+            gre_dates.append(date(*gre.datetuple()))
+    return gre_dates
